+++ packages/context-engine/src/processors/MessageContent.ts
