+++ packages/context-engine/src/processors/PlaceholderVariables.ts
@@ -8,16 +8,16 @@ const log = debug('context-engine:processor:PlaceholderVariablesProcessor');
 const placeholderVariablesRegex = /{{(.*?)}}/g;
 
 export interface PlaceholderVariablesConfig {
-  /** 递归解析深度，默认为 2 */
+  /** Recursive parsing depth, default is 2 */
   depth?: number;
-  /** 变量生成器映射，key 为变量名，value 为生成函数 */
+  /** Variable generators mapping, key is variable name, value is generator function */
   variableGenerators: Record<string, () => string>;
 }
 
 /**
- * 从文本中提取所有 {{variable}} 占位符的变量名
- * @param text 包含模板变量的字符串
- * @returns 变量名数组，如 ['date', 'nickname']
+ * Extract all {{variable}} placeholder variable names from text
+ * @param text String containing template variables
+ * @returns Array of variable names, e.g. ['date', 'nickname']
  */
 const extractPlaceholderVariables = (text: string): string[] => {
   const matches = [...text.matchAll(placeholderVariablesRegex)];
@@ -25,11 +25,11 @@ const extractPlaceholderVariables = (text: string): string[] => {
 };
 
 /**
- * 将模板变量替换为实际值，并支持递归解析嵌套变量
- * @param text - 含变量的原始文本
- * @param variableGenerators - 变量生成器映射
- * @param depth - 递归深度，默认 2，设置更高可支持 {{text}} 中的 {{date}} 等
- * @returns 替换后的文本
+ * Replace template variables with actual values, supporting recursive parsing of nested variables
+ * @param text - Original text containing variables
+ * @param variableGenerators - Variable generators mapping
+ * @param depth - Recursive depth, default 2, set higher to support {{date}} within {{text}}
+ * @returns Text with variables replaced
  */
 export const parsePlaceholderVariables = (
   text: string,
@@ -38,7 +38,7 @@ export const parsePlaceholderVariables = (
 ): string => {
   let result = text;
 
-  // 递归解析，用于处理如 {{text}} 存在额外预设变量
+  // Recursive parsing to handle cases like {{text}} containing additional preset variables
   for (let i = 0; i < depth; i++) {
     try {
       const extractedVariables = extractPlaceholderVariables(result);
@@ -48,10 +48,10 @@ export const parsePlaceholderVariables = (
           .filter(([, value]) => value !== undefined),
       );
 
-      // 只有当有可用变量时才进行替换
+      // Only perform replacement when there are available variables
       if (Object.keys(availableVariables).length === 0) break;
 
-      // 逐个替换变量，避免 lodash template 对未定义变量的错误处理
+      // Replace variables one by one to avoid lodash template's error handling for undefined variables
       let tempResult = result;
       for (const [key, value] of Object.entries(availableVariables)) {
         const regex = new RegExp(
@@ -73,11 +73,11 @@ export const parsePlaceholderVariables = (
 };
 
 /**
- * 解析消息内容，替换占位符变量
- * @param messages 原始消息数组
- * @param variableGenerators 变量生成器映射
- * @param depth 递归解析深度，默认为 2
- * @returns 处理后的消息数组
+ * Parse message content and replace placeholder variables
+ * @param messages Original messages array
+ * @param variableGenerators Variable generators mapping
+ * @param depth Recursive parsing depth, default is 2
+ * @returns Processed messages array
  */
 export const parsePlaceholderVariablesMessages = (
   messages: any[],
@@ -89,12 +89,12 @@ export const parsePlaceholderVariablesMessages = (
 
     const { content } = message;
 
-    // 字符串类型直接处理
+    // Handle string type directly
     if (typeof content === 'string') {
       return { ...message, content: parsePlaceholderVariables(content, variableGenerators, depth) };
     }
 
-    // 数组类型处理其中的 text 元素
+    // Handle array type by processing text elements
     if (Array.isArray(content)) {
       return {
         ...message,
@@ -111,7 +111,7 @@ export const parsePlaceholderVariablesMessages = (
 
 /**
  * PlaceholderVariables Processor
- * 负责处理消息中的占位符变量替换
+ * Responsible for handling placeholder variable replacement in messages
  */
 export class PlaceholderVariablesProcessor extends BaseProcessor {
   readonly name = 'PlaceholderVariablesProcessor';
@@ -168,7 +168,7 @@ export class PlaceholderVariablesProcessor extends BaseProcessor {
 
     const { content } = message;
 
-    // 字符串类型直接处理
+    // Handle string type directly
     if (typeof content === 'string') {
       return {
         ...message,
@@ -176,7 +176,7 @@ export class PlaceholderVariablesProcessor extends BaseProcessor {
       };
     }
 
-    // 数组类型处理其中的 text 元素
+    // Handle array type by processing text elements
     if (Array.isArray(content)) {
       return {
         ...message,
