+++ packages/context-engine/src/processors/index.ts
@@ -1,14 +1,14 @@
 // Transformer processors
-export { MessageCleanupProcessor } from './MessageCleanupProcessor';
-export { MessageContentProcessor } from './MessageContentProcessor';
-export { PlaceholderVariablesProcessor } from './PlaceholderVariablesProcessor';
-export { ToolCallProcessor } from './ToolCallProcessor';
+export { MessageCleanupProcessor } from './MessageCleanup';
+export { MessageContentProcessor } from './MessageContent';
+export { PlaceholderVariablesProcessor } from './PlaceholderVariables';
+export { ToolCallProcessor } from './ToolCall';
 export { ToolMessageReorder } from './ToolMessageReorder';
 
 // Validator processors
 export { ModelCapabilityValidator } from './ModelCapabilityValidator';
 
 // Re-export types
-export type { MessageContentConfig, UserMessageContentPart } from './MessageContentProcessor';
-export type { PlaceholderVariablesConfig } from './PlaceholderVariablesProcessor';
-export type { ToolCallConfig } from './ToolCallProcessor';
+export type { MessageContentConfig, UserMessageContentPart } from './MessageContent';
+export type { PlaceholderVariablesConfig } from './PlaceholderVariables';
+export type { ToolCallConfig } from './ToolCall';
